Add migrate command
This command has a --truncate flag that can be used to truncate the
destination tables prior to doing the migration.

The migration is optimized for tables that have an "id" column:

1. Select all the ids that already exist in MySQL for that table
2. Select the data for ids which do *not* exist in MySQL from PostgreSQL
3. Insert that data into MySQL, doing one INSERT per row from PostgreSQL

In the worst-case scenario I think the above will do one SELECT in MySQL
and one in PostgreSQL per table plus one INSERT into MySQL per row that
needs to be migrated.

The above could be optimized further to do batch INSERTs into MySQL (say
in batches of 10 or 50), but I'll leave that for after we actually test
this and see how it does on a production-like database.

For tables that don't have an "id" column we have do do a full-row
comparison in MySQL. These are typically join tables, so the data can be
compared exactly (whereas timestamps cannot be, due to differences in
precision).

The steps for tables that don't have an "id" column:

1. Select all rows in PostgreSQL for the table
2. Determine if the row exists in MySQL by doing `SELECT EXISTS` with
   all of the values for the row
3. If the row does not exist in MySQL, INSERT it

The worst-case scenario for the above is one SELECT in PosgreSQL for
each table plus one SELECT *and* one INSERT per row that needs to be
migrated in MySQL. MariaDB does support doing INSERT ON DUPLICATE
KEY UPDATE, but that would end up doing updates to each row already
present every time the migration is run. I chose to go the non-update
route as reads are cheaper than writes.

[#142280047]