Reconnect to mysql in mysqlrunner
I was seeing multiple 'no database selected' errors even though there's
an exec at the top of the test file to use the runner's database.

Using the original dbconn to just create the database, close that
dbconn, and open a new one with the updated config seems to work more
reliably.